# openapi.json
{
  "components": {
    "schemas": {
...
      "Priority": {
        "enum": [
          "p0",
          "p1",
          "p2",
          "p3"
        ],
        "title": "Priority",
        "type": "string"
      },
      "Problem": {
        "description": "RFC 9457 ``application/problem+json`` — **every** error response.\n\nDeclared as a model, and attached to the routers, for a reason that is easy to\nmiss: the exception handlers are invisible to FastAPI's schema generation, so\nwithout this the OpenAPI document would describe only the happy paths. An\nintegrator reading ``/docs`` would then have to *discover* the error shape by\ncausing errors, which is precisely the \"first thing an integrator hits\" §8.6\nis about. It also means the shape lands in API-5's snapshot, so changing it\nbecomes a visible contract change.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldProblem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Present on a 422; one entry per rejected field.",
            "title": "Errors"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "Human-readable, and names the next step.",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "Stable URI identifying the error class. **Match on this**, never on ``title`` — the title is human-readable prose and may be reworded.",
            "examples": [
              "https://relay.internal/problems/not_found"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "TicketStatus": {
        "description": "TKT-3 · **frozen from release** — these values ship in API responses.",
        "enum": [
          "todo",
          "in_progress",
          "in_review",
          "done",
          "blocked",
          "wont_fix"
        ],
        "title": "TicketStatus",
        "type": "string"
      },
...
      "relay__api__v1__tickets__TicketPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/relay__api__v1__tickets__TicketResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "TicketPage",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Relay",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/tickets": {
      "get": {
        "description": "Newest-updated first, keyset-paged (§8.3).\n\n``updated_since`` is what a polling consumer uses, and it is **inclusive** —\nan exclusive bound drops exactly the rows whose timestamp equals the\nwatermark the caller stored. Cursor rather than offset because the ordering\nkey changes while somebody pages, and offset both skips and repeats rows.\n\nIt is typed as a string rather than a ``datetime`` on purpose; see\n:func:`_instant`.",
        "operationId": "list_tickets_api_v1_tickets_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/TicketStatus"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "priority",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/Priority"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          },
          {
            "in": "query",
            "name": "assignee",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assignee"
            }
          },
          {
            "in": "query",
            "name": "label",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Label"
            }
          },
          {
            "in": "query",
            "name": "iteration",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Iteration"
            }
          },
          {
            "in": "query",
            "name": "updated_since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Updated Since"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/relay__api__v1__tickets__TicketPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Refused — e.g. a tenant_id in the request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Missing, expired or revoked token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Out of scope for this token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No such resource — or another tenant's"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "rev mismatch; the body carries the current rev"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation failed; see errors[]"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Rate limited; see Retry-After"
          }
        },
        "summary": "List Tickets",
        "tags": [
          "tickets (v1)"
        ]
      },
      "post": {
        "description": "File a ticket. **201**, or **200** when it already existed.\n\nThree outcomes, and the distinction between the last two is the whole of\n§8.4's \"both defences are needed\":\n\n* new ticket → 201 with ``Location``;\n* the same ``Idempotency-Key`` replayed → the *first* response, verbatim.\n  This is the network defence: a lost response, a proxy retry;\n* a matching ``external_ref`` → **200** with the ticket that already exists.\n  This is the upstream defence: a user clicking submit three times, or a\n  compensating job re-running. It is not an error, because the caller needs\n  to be told which ticket to look at rather than given something to retry.",
        "operationId": "create_ticket_api_v1_tickets_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/relay__api__v1__tickets__CreateTicketPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/relay__api__v1__tickets__TicketResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Refused — e.g. a tenant_id in the request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Missing, expired or revoked token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Out of scope for this token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No such resource — or another tenant's"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "rev mismatch; the body carries the current rev"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation failed; see errors[]"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Rate limited; see Retry-After"
          }
        },
        "summary": "Create Ticket",
        "tags": [
          "tickets (v1)"
        ]
      }
    },
    "/api/v1/tickets/{key}": {
      "get": {
        "description": "Detail. This is also F-6 ①'s polling endpoint: the feedback consumer reads\n``status`` and ``updated_at`` from here and shows progress to the submitter.\n\n⚠️ Internal comments are **not** hidden from this token — ``/comments`` is\nreadable with ``tickets:read``. \"Do not show internal comments to the\nsubmitter\" is a constraint on the *consumer* and belongs in the integration\nguide; expecting the API surface to enforce it would be a misreading.",
        "operationId": "get_ticket_api_v1_tickets__key__get",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/relay__api__v1__tickets__TicketResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Refused — e.g. a tenant_id in the request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Missing, expired or revoked token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Out of scope for this token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No such resource — or another tenant's"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "rev mismatch; the body carries the current rev"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation failed; see errors[]"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Rate limited; see Retry-After"
          }
        },
        "summary": "Get Ticket",
        "tags": [
          "tickets (v1)"
        ]
      },
      "patch": {
        "description": "Partial update. ``If-Match: <rev>`` required; 409 carries the current rev.\n\nNot idempotency-keyed, and that is not an omission: ``rev`` already makes a\nreplayed PATCH safe — the second attempt carries a stale rev and is refused\nwith the current one, which is exactly the information the client needs.",
        "operationId": "update_ticket_api_v1_tickets__key__patch",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/relay__api__v1__tickets__UpdateTicketPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/relay__api__v1__tickets__TicketResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Refused — e.g. a tenant_id in the request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Missing, expired or revoked token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Out of scope for this token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No such resource — or another tenant's"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "rev mismatch; the body carries the current rev"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation failed; see errors[]"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Rate limited; see Retry-After"
          }
        },
        "summary": "Update Ticket",
        "tags": [
          "tickets (v1)"
        ]
      }
    },
...
    "/api/v1/tickets/{key}/history": {
      "get": {
        "operationId": "ticket_history_api_v1_tickets__key__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HistoryResponse"
                  },
                  "title": "Response Ticket History Api V1 Tickets  Key  History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Refused — e.g. a tenant_id in the request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Missing, expired or revoked token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Out of scope for this token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No such resource — or another tenant's"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "rev mismatch; the body carries the current rev"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation failed; see errors[]"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Rate limited; see Retry-After"
          }
        },
        "summary": "Ticket History",
        "tags": [
          "tickets (v1)"
        ]
      }
    },
...
  }
}